Various refactoring
