+++ src/main/java/com/jagrosh/jmusicbot/commands/PlayCommand.java
@@ -1,0 +1,121 @@
+/*
+ * Copyright 2018 John Grosh <john.a.grosh@gmail.com>.
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *      http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+package com.jagrosh.jmusicbot.commands;
+
+import com.jagrosh.jdautilities.command.CommandEvent;
+import com.jagrosh.jmusicbot.Bot;
+import com.jagrosh.jmusicbot.audio.AudioHandler;
+import com.jagrosh.jmusicbot.audio.QueuedTrack;
+import com.jagrosh.jmusicbot.commands.DJCommand;
+import com.jagrosh.jmusicbot.utils.FormatUtil;
+import com.sedmelluq.discord.lavaplayer.player.AudioLoadResultHandler;
+import com.sedmelluq.discord.lavaplayer.tools.FriendlyException;
+import com.sedmelluq.discord.lavaplayer.track.AudioPlaylist;
+import com.sedmelluq.discord.lavaplayer.track.AudioTrack;
+import net.dv8tion.jda.core.entities.Message;
+
+/**
+ *
+ * @author Daniel Liljeberg (liljebergxyz@gmail.com)
+ */
+public abstract class PlayCommand extends DJCommand
+{
+    protected final String loadingEmoji;
+
+    public PlayCommand(Bot bot, String loadingEmoji)
+    {
+        super(bot);
+        this.loadingEmoji = loadingEmoji;
+        this.arguments = "<title|URL>";
+        this.beListening = true;
+        this.bePlaying = false;
+    }
+
+    public class PlayResultHandler implements AudioLoadResultHandler
+    {
+        protected final Message m;
+        protected final CommandEvent event;
+        private final boolean ytsearch;
+        
+        public PlayResultHandler(Message m, CommandEvent event, boolean ytsearch)
+        {
+            this.m = m;
+            this.event = event;
+            this.ytsearch = ytsearch;
+        }
+        
+        protected void loadSingle(AudioTrack track)
+        {
+            AudioHandler handler = (AudioHandler)event.getGuild().getAudioManager().getSendingHandler();
+            handler.addTrackToFront(new QueuedTrack(track, event.getAuthor()));
+            String addMsg = FormatUtil.filter(event.getClient().getSuccess()+" Added **"+track.getInfo().title
+                    +"** (`"+FormatUtil.formatTime(track.getDuration())+"`) to begin playing");
+            m.editMessage(addMsg).queue();
+            handler.getPlayer().stopTrack();
+        }
+        
+        private Boolean isTrackShortEnough(AudioTrack track)
+        {
+            if(bot.getConfig().isTooLong(track))
+            {
+                m.editMessage(FormatUtil.filter(event.getClient().getWarning()+" This track (**"+track.getInfo().title+"**) is longer than the allowed maximum: `"
+                        +FormatUtil.formatTime(track.getDuration())+"` > `"+FormatUtil.formatTime(bot.getConfig().getMaxSeconds()*1000)+"`")).queue();
+                return false;
+            }
+
+            return true;
+        }
+        
+        @Override
+        public void trackLoaded(AudioTrack track)
+        {
+            if (isTrackShortEnough(track)) {
+                loadSingle(track);
+            }
+        }
+
+        @Override
+        public void playlistLoaded(AudioPlaylist playlist)
+        {
+            AudioTrack single;
+            if(playlist.getTracks().size()==1 || playlist.isSearchResult())
+                single = playlist.getSelectedTrack()==null ? playlist.getTracks().get(0) : playlist.getSelectedTrack();
+            else if (playlist.getSelectedTrack()!=null)
+                single = playlist.getSelectedTrack();
+            else
+                single = playlist.getTracks().get(0);
+            loadSingle(single);
+        }
+
+        @Override
+        public void noMatches()
+        {
+            if(ytsearch)
+                m.editMessage(FormatUtil.filter(event.getClient().getWarning()+" No results found for `"+event.getArgs()+"`.")).queue();
+            else
+                bot.getPlayerManager().loadItemOrdered(event.getGuild(), "ytsearch:"+event.getArgs(), new PlayResultHandler(m,event,true));
+        }
+
+        @Override
+        public void loadFailed(FriendlyException throwable)
+        {
+            if(throwable.severity==FriendlyException.Severity.COMMON)
+                m.editMessage(event.getClient().getError()+" Error loading: "+throwable.getMessage()).queue();
+            else
+                m.editMessage(event.getClient().getError()+" Error loading track.").queue();
+        }
+    }
+}
