+++ src/main/java/com/jagrosh/jmusicbot/commands/dj/PlaynowCmd.java
@@ -18,32 +18,19 @@ package com.jagrosh.jmusicbot.commands.dj;
 import com.jagrosh.jdautilities.command.CommandEvent;
 import com.jagrosh.jmusicbot.Bot;
 import com.jagrosh.jmusicbot.audio.AudioHandler;
-import com.jagrosh.jmusicbot.audio.QueuedTrack;
-import com.jagrosh.jmusicbot.commands.DJCommand;
-import com.jagrosh.jmusicbot.utils.FormatUtil;
-import com.sedmelluq.discord.lavaplayer.player.AudioLoadResultHandler;
-import com.sedmelluq.discord.lavaplayer.tools.FriendlyException;
-import com.sedmelluq.discord.lavaplayer.track.AudioPlaylist;
-import com.sedmelluq.discord.lavaplayer.track.AudioTrack;
-import net.dv8tion.jda.core.entities.Message;
+import com.jagrosh.jmusicbot.commands.PlayCommand;
 
 /**
  *
  * @author Daniel Liljeberg (liljebergxyz@gmail.com)
  */
-public class PlaynowCmd extends DJCommand
+public class PlaynowCmd extends PlayCommand
 {
-    private final String loadingEmoji;
-    
     public PlaynowCmd(Bot bot, String loadingEmoji)
     {
-        super(bot);
-        this.loadingEmoji = loadingEmoji;
+        super(bot, loadingEmoji);
         this.name = "playnow";
-        this.arguments = "<title|URL>";
         this.help = "plays a single song now";
-        this.beListening = true;
-        this.bePlaying = false;
     }
     
     @Override
@@ -59,73 +46,6 @@ public class PlaynowCmd extends DJCommand
         String args = event.getArgs().startsWith("<") && event.getArgs().endsWith(">") 
                 ? event.getArgs().substring(1,event.getArgs().length()-1) 
                 : event.getArgs().isEmpty() ? event.getMessage().getAttachments().get(0).getUrl() : event.getArgs();
-        event.reply(loadingEmoji+" Loading... `["+args+"]`", m -> bot.getPlayerManager().loadItemOrdered(event.getGuild(), args, new ResultHandler(m,event,false)));
-    }
-    
-    private class ResultHandler implements AudioLoadResultHandler
-    {
-        private final Message m;
-        private final CommandEvent event;
-        private final boolean ytsearch;
-        
-        private ResultHandler(Message m, CommandEvent event, boolean ytsearch)
-        {
-            this.m = m;
-            this.event = event;
-            this.ytsearch = ytsearch;
-        }
-        
-        private void loadSingle(AudioTrack track)
-        {
-            if(bot.getConfig().isTooLong(track))
-            {
-                m.editMessage(FormatUtil.filter(event.getClient().getWarning()+" This track (**"+track.getInfo().title+"**) is longer than the allowed maximum: `"
-                        +FormatUtil.formatTime(track.getDuration())+"` > `"+FormatUtil.formatTime(bot.getConfig().getMaxSeconds()*1000)+"`")).queue();
-                return;
-            }
-            AudioHandler handler = (AudioHandler)event.getGuild().getAudioManager().getSendingHandler();
-            handler.addTrackToFront(new QueuedTrack(track, event.getAuthor()));
-            String addMsg = FormatUtil.filter(event.getClient().getSuccess()+" Added **"+track.getInfo().title
-                    +"** (`"+FormatUtil.formatTime(track.getDuration())+"`) to begin playing");
-            m.editMessage(addMsg).queue();
-            handler.getPlayer().stopTrack();
-        }
-        
-        @Override
-        public void trackLoaded(AudioTrack track)
-        {
-            loadSingle(track);
-        }
-
-        @Override
-        public void playlistLoaded(AudioPlaylist playlist)
-        {
-            AudioTrack single;
-            if(playlist.getTracks().size()==1 || playlist.isSearchResult())
-                single = playlist.getSelectedTrack()==null ? playlist.getTracks().get(0) : playlist.getSelectedTrack();
-            else if (playlist.getSelectedTrack()!=null)
-                single = playlist.getSelectedTrack();
-            else
-                single = playlist.getTracks().get(0);
-            loadSingle(single);
-        }
-
-        @Override
-        public void noMatches()
-        {
-            if(ytsearch)
-                m.editMessage(FormatUtil.filter(event.getClient().getWarning()+" No results found for `"+event.getArgs()+"`.")).queue();
-            else
-                bot.getPlayerManager().loadItemOrdered(event.getGuild(), "ytsearch:"+event.getArgs(), new ResultHandler(m,event,true));
-        }
-
-        @Override
-        public void loadFailed(FriendlyException throwable)
-        {
-            if(throwable.severity==FriendlyException.Severity.COMMON)
-                m.editMessage(event.getClient().getError()+" Error loading: "+throwable.getMessage()).queue();
-            else
-                m.editMessage(event.getClient().getError()+" Error loading track.").queue();
-        }
+        event.reply(loadingEmoji+" Loading... `["+args+"]`", m -> bot.getPlayerManager().loadItemOrdered(event.getGuild(), args, new PlayResultHandler(m,event,false)));
     }
 }
